requirements.txt
opencv-python-headless
fastapi
uvicorn
pillow
uvloop
loguru